split out sessions and topics
