+++ src/app/nav/nav.component.html
@@ -34,7 +34,7 @@
         </a>
         <ul *dropdownMenu class="dropdown-menu nav-background navbar-dark">
             <li><a pageScroll class="dropdown-item" routerLink="/resources" href="#tools">The Tools</a></li>
-            <li><a pageScroll class="dropdown-item" routerLink="/resources" href="#list">Monthly Slides &amp; Worksheets</a></li>
+            <li><a pageScroll class="dropdown-item" routerLink="/resources" href="#list">Slides &amp; Worksheets</a></li>
             <li><a pageScroll class="dropdown-item" routerLink="/resources" href="#resources">Learning Resources</a></li>
         </ul>
       </li>
