+++ src/app/resources/resources.component.html
@@ -17,18 +17,18 @@
 </section>
 <section id="list">
   <div class="list">
-    <h3>Monthly Slides &amp; Worksheets</h3>
+    <h3>Slides &amp; Worksheets</h3>
     <table class="tocTable">
-      <tr *ngFor="let session of sessions">
-          <td style="text-align: left;">{{session.date | date:'LLLL'}}: {{session.sessionName}}</td>
-          <ng-template [ngIf]="!session.areSessionLinksEnabled" [ngIfElse]="available">
+      <tr *ngFor="let topic of topics">
+          <td style="text-align: left;">{{topic.name}}</td>
+          <ng-template [ngIf]="!topic.areLinksEnabled" [ngIfElse]="available">
               <td><font style="color: #999">slides</font></td>
               <td><font style="color: #999">worksheet</font></td>
           </ng-template>
 
           <ng-template #available>
-              <td><a href="{{session.slidesLink}}" class="listlink" target="_blank">slides</a></td>
-              <td><a href="{{session.worksheetLink}}" class="listlink" target="_blank">worksheet</a></td>
+              <td><a href="{{topic.slidesLink}}" class="listlink" target="_blank">slides</a></td>
+              <td><a href="{{topic.worksheetLink}}" class="listlink" target="_blank">worksheet</a></td>
           </ng-template>
       </tr>
     </table>
