+++ src/data/heroes.ts
@@ -1,0 +1,137 @@
+export interface Hero {
+    name: string;
+    title: string;
+    img: string;
+    twitter: string;
+}
+
+export const LEADERS: Hero[] = [{
+    name: 'Jennifer Wadella',
+    title: 'Founder',
+    img: 'assets/images/mentors/jennifer.png',
+    twitter: 'likeOMGitsFEDAY'
+}, {
+    name: 'Sarah Duitsman',
+    title: 'Program Director',
+    img: 'assets/images/mentors/sarah.png',
+    twitter: 'skduits'
+}, {
+    name: 'Rachel Hagan',
+    title: 'Co Director',
+    img: 'assets/images/mentors/rachelh.jpg',
+    twitter: 'rhagan21'
+}, {
+    name: 'Kayla Hennegin',
+    title: 'Curriculum Director',
+    img: 'assets/images/mentors/kayla.png',
+    twitter: 'codecocktailskc'
+}, {
+    name: 'Gabi Dombrowski',
+    title: 'Mentor Director',
+    img: 'assets/images/mentors/gabi.jpg',
+    twitter: 'gabdom'
+}, {
+    name: 'Alisa Duncan',
+    title: 'Technical Materials Director',
+    img: 'assets/images/mentors/alisa.jpg',
+    twitter: 'AlisaDuncan'
+}, {
+    name: 'Melissa Arroyo',
+    title: 'Marketing Director',
+    img: 'assets/images/mentors/melissa.jpg',
+    twitter: 'codecocktailskc'
+}, {
+    name: 'Kallie Riff',
+    title: 'Presentation Director',
+    img: 'assets/images/mentors/kallie.jpg',
+    twitter: 'KallieRiff'
+}];
+
+
+export const MENTORS: Hero[] = [{
+    name: 'Tamara Copple',
+    title: 'Mentor',
+    img: 'assets/images/mentors/tamara.png',
+    twitter: 'Ceiligirl72'
+}, {
+    name: 'Amy Norris',
+    title: 'Mentor',
+    img: 'assets/images/mentors/amy.jpg',
+    twitter: 'amynorris'
+}, {
+    name: 'Rachel Hathaway',
+    title: 'Mentor',
+    img: 'assets/images/mentors/rachel.jpg',
+    twitter: 'codecocktailskc'
+}, {
+    name: 'Muriel Green',
+    title: 'Mentor',
+    img: 'assets/images/mentors/muriel.jpg',
+    twitter: 'murlgreen1'
+}, {
+    name: 'Annie Chan',
+    title: 'Mentor',
+    img: 'assets/images/mentors/annie.jpg',
+    twitter: 'codecocktailskc'
+}, {
+    name: 'Heather Pekrul',
+    title: 'Mentor',
+    img: 'assets/images/mentors/heatherp.jpg',
+    twitter: 'heatherpekrul'
+}, {
+    name: 'Maroua Jawadi',
+    title: 'Mentor',
+    img: 'assets/images/mentors/maroua.jpg',
+    twitter: 'codecocktailskc'
+}, {
+    name: 'Stephanie Sammons',
+    title: 'Mentor',
+    img: 'assets/images/mentors/stephanie.jpg',
+    twitter: 'codecocktailskc'
+}, {
+    name: 'Rhia Dixon',
+    title: 'Mentor',
+    img: 'assets/images/mentors/rhia.jpg',
+    twitter: 'TheFakeRiRi'
+}, {
+    name: 'Stacey Adams',
+    title: 'Mentor',
+    img: 'assets/images/mentors/stacey.jpg',
+    twitter: 'codecocktailskc'
+}, {
+    name: 'Melanie Garvey',
+    title: 'Mentor',
+    img: 'assets/images/mentors/mel.jpg',
+    twitter: 'MelanieGarvey'
+}, {
+    name: 'McKenzie Hine',
+    title: 'Mentor',
+    img: 'assets/images/mentors/mckenzie.jpg',
+    twitter: 'mckenz_h'
+}, {
+    name: 'Jenn Bono',
+    title: 'Mentor',
+    img: 'assets/images/mentors/jenn.jpg',
+    twitter: 'jenn_bono24'
+}, {
+    name: 'Erin Wilson',
+    title: 'Mentor',
+    img: 'assets/images/mentors/erin.jpg',
+    twitter: 'erin_wilson92'
+}, {
+    name: 'Gabby Ortman',
+    title: 'Mentor',
+    img: 'assets/images/mentors/gabby.jpg',
+    twitter: 'gabbyortman'
+}, {
+    name: 'Jessica Kincaid',
+    title: 'Mentor',
+    img: 'assets/images/mentors/jessica.jpg',
+    twitter: 'jessica_k1ncaid'
+}, {
+    name: 'Gloria Highley',
+    title: 'Mentor',
+    img: 'assets/images/mentors/gloria.jpg',
+    twitter: 'gloriahigley'
+}];
+
