+++ src/data/index.ts
@@ -1,0 +1,3 @@
+export * from './about';
+export * from './heroes';
+export * from './sessions';
